requirements.txt
# Library for reading and parsing PDFs
pypdf

# Python wrapper for the Tesseract OCR engine
pytesseract

# Streamlit for the web application
streamlit

# Other standard libraries
Pillow